requirements.txt
selenium
pillow
Beautifulsoup4
pandas
openpyxl
xlrd